+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.18.1 (TBD)
+- Added mastery for Ticonderogun
 - Added 1 new synergy
 - Fixed dummy items from being able to appear in synergy chests
 - Fixed Bubblebeam permanently changing the color of some projectiles fired by enemies even when they aren't enbubbled